backend_django/requirements.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.0
python-decouple>=3.8
Pillow>=10.0
whitenoise>=6.0
requests>=2.31.0
gunicorn>=21.0
dj-database-url>=2.0
psycopg2-binary>=2.9
